Choice Picker
In this code, we define a function random_picker that takes a list of options as an argument, selects a random option from the list, and prints the result.

The user is prompted to enter a comma-separated list of options. The input is then split into a list using the split(',') function, and this list of options is passed to the random_picker function.

For example, if the user enters "Option A,Option B,Option C", the program will randomly select one of these options and display the result.